Update patch set 1
Patch Set 1:

I know it's work but let's fix style a bit while we're touching the code.

Regarding line length, should we raise the checkpatch line length limit 100?
In this case the indent in function continuation was incorrect (we normally align to open parent so the argument continuation lines will be <80) but there are still many places in mckernel code where we're far from 80 and these messages are annoying...

Patch-set: 1
CC: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>