Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > I'm not sure about moving code to arch-specific code, it's not like they conflict?
> > Actually, x86 has both open and openat so you would need to duplicate openat with your logic.
> 
> Thanks for pointing out! It's changed to move only sys_open and xpmem_open to x86_64.

That's not the only reason I said that though - for me anything that doesn't explicitly conflict should be in the common code, with as little as possible in the arch directories.
The reason is that if we add a new arch someday or if the syscalls expected of arm64 ever change, it will be more work to add the function back, when there was no reason to take it away in the first place.

We should really limit arch to the minimum required.
(on the other hand, I also think kernel/syscall.c is too big, but splitting it would be something else to do if we ever run out of things to cleanup...!)


> > Also, let's fix a few of the checkpatch warnings please - some like pointer on the wrong side, mixed spaces etc should be fixed.
> > (not a checkpatch warning, but also remove extern for function declarations!)
> 
> It's done.

Thanks!

Patch-set: 7
Reviewer: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>