Update patch set 1
Patch Set 1:

same comment as other commit modifying kernel/gencore.c: this file has been deleted and unless proven it is useful I do not plan to add it back with the huge arm patch; please do not waste time on it.

There are equivalent arch-specific files, it might make sense to add a common file again but right now it is not used.

Patch-set: 1
CC: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>