Update patch set 3
Patch Set 3:

> > I don't understand why this helps. This definitely changes timing a bit so might make tests work by chance, but places that update proc->parent / look at proc->status do not coherently have this lock as far as I can see?
> 
> The problem of updating proc-> status and proc-> parent as a pair is not a proc-> status alone.

Yes, but other places do not keep the lock for both either, e.g. for parent update:
 - clone() probably doesn't count for this race as process is being created
 - wait_proc() on parent will take the child's update_lock briefly, then release it after updating only ->parent
 - ptrace_detach_thread does not take even the tracee proc's update lock, only its children lock...

Likewise for status/parent lookup:
 - ptrace_detach_thread makes similar decision (to call finalize_process or not) without taking the update_lock as a reader
 - actually almost nothing takes this read lock for ->status...

I'm not blaming you for the inconsistent state we're in, but I really think that your bug here is resolved because of a single particular access pattern and that it's not a general solution. If you explain what threads are racing where (which functions are involved) I might agree we need to add that lock here and start locking more seriously; or try to come up with something else.

Patch-set: 3