Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> No clue how links work in the editor you use, but this should work for github after inspecting the generated html there... Hopefully.
> 
> re-filename you didn't answer my comment so being pushy instead :p I'd go as far as drop the .md actually, it's not required for github, but I guess it might help with editor/file browsers recognizing files as markdown?

It worked in the markdown editor I used and yea after noticing it I also had the background link fixed in my local repo (although I just simply changed it to background). 
As for the filename, please don't rename it though, I would prefer if it stays clear that this is a markdown file and not a regular readme txt.

Patch-set: 1
CC: Gerrit User 1015205 <1015205@d5d70762-12d0-45a1-890d-524b12d3f735>