Update patch set 2
Abandoned

dropping this, I'm not sure why this fails vmsplice01 to be honest, but I had read the code of vm_insert_mixed too fast -- there's an arch option "PTE_SPECIAL" that x86/arm have that make this not use insert_page when we would. Can't remember why I even wanted to do that in the first place, so abandon it is.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon