Update patch set 1
Patch Set 1:

> Since there was a code fix to return -1 when ihk_mc_get_processor_id() was executed on Linux on x86 implementation, it was also implemented by arm64, but was it unnecessary?

Ah, I see why you wanted to group this together.
It can be useful for debugging purposes but it does not have to be in the same commit as currently nothing will call into the mckernel text from linux.

I'd still prefer this to be in a separate commit but if you really want them to be together please include a comment in the commit message as I had originally suggested.

Patch-set: 1
Reviewer: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>