Update patch set 1
Patch Set 1:

Well, that didn't work so well...

Past the obvious build problem; I have one problem with this as it has been annoying me on arm: this makes the summary output of CFLAGS and build commands really long.

Here's the ones that are left according to a quick grep:
      6 POSTK_DEBUG_TEMP_FIX_11
      3 POSTK_DEBUG_TEMP_FIX_21
     12 POSTK_DEBUG_TEMP_FIX_29
      4 POSTK_DEBUG_TEMP_FIX_30
      8 POSTK_DEBUG_TEMP_FIX_33
      4 POSTK_DEBUG_TEMP_FIX_37
     16 POSTK_DEBUG_TEMP_FIX_39
      3 POSTK_DEBUG_TEMP_FIX_49
      8 POSTK_DEBUG_TEMP_FIX_63
      1 POSTK_DEBUG_ARCH_DEP_10
      4 POSTK_DEBUG_ARCH_DEP_100
      1 POSTK_DEBUG_ARCH_DEP_105
      2 POSTK_DEBUG_ARCH_DEP_110
     14 POSTK_DEBUG_ARCH_DEP_12
      2 POSTK_DEBUG_ARCH_DEP_18
      2 POSTK_DEBUG_ARCH_DEP_21
      9 POSTK_DEBUG_ARCH_DEP_29
     63 POSTK_DEBUG_ARCH_DEP_34
      6 POSTK_DEBUG_ARCH_DEP_38
     36 POSTK_DEBUG_ARCH_DEP_40
      9 POSTK_DEBUG_ARCH_DEP_42
      1 POSTK_DEBUG_ARCH_DEP_43
      3 POSTK_DEBUG_ARCH_DEP_50
     12 POSTK_DEBUG_ARCH_DEP_52
      2 POSTK_DEBUG_ARCH_DEP_53
      5 POSTK_DEBUG_ARCH_DEP_56
      3 POSTK_DEBUG_ARCH_DEP_60
      2 POSTK_DEBUG_ARCH_DEP_63
      3 POSTK_DEBUG_ARCH_DEP_66
      2 POSTK_DEBUG_ARCH_DEP_70
      6 POSTK_DEBUG_ARCH_DEP_75
      4 POSTK_DEBUG_ARCH_DEP_77
     27 POSTK_DEBUG_ARCH_DEP_8
      2 POSTK_DEBUG_ARCH_DEP_9
      4 POSTK_DEBUG_ARCH_DEP_90

Can we maybe get rid of the ones with <=3 occurences quickly and only hand-pick the ones that are left? It'll make the cmakefile a bit uglier but it'll also serve to make sure we don't add new ones.

Patch-set: 1
CC: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>