Update patch set 2
Patch Set 2:

> I was forgotten that hwid is 32 bit configuration.
> If you compare with the setting contents of already existing cpu_logical_map,
> what you want to do can be realized.
> However, since cpu_logical_map is an array, I thought that array searching occurred every time ihk_mc_get_processor_id() was executed has a bitmap, thinking that performance is not good.

I am sorry I do not understand what you said here, and I am not sure "what I want to do" refers to either (it might help to reply to one of the three points specifically)

I think we can still use a bitmap if you want to though, that is not the problem is it?

> Since we are asking that only do_futex() is necessary to consider what is executed on HOST, we do not always check it with ihk_mc_get_processor_id(), but how about checking only what you call with do_futex()?

I thought the point of this patch is to make ihk_mc_get_processor_id() fail if it is called from a linux core, because from the moment we map mckernel text section on linux we will be able to start doing this (and e.g. the hfi driver offloading we have in the ofp branch does this so might call into code that can behave differently if processor_id is -1)

If this is about do_futex() it might be enough but I still am not sure what you said, please re-explain first

Patch-set: 2
Reviewer: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>