Update patch set 2
Patch Set 2:

> I am sorry I do not understand what you said here, and I am not sure "what I want to do" refers to either (it might help to reply to one of the three points specifically)
> 
> I think we can still use a bitmap if you want to though, that is not the problem is it?

When implementing with bitmap, to cover all 32 bit expressions, an area of 4294967296 bits is required.
I believe that is not realistic.
Although it can be realized by using cpu_logical_map, there is a concern that performance will be affected.

> I thought the point of this patch is to make ihk_mc_get_processor_id() fail if it is called from a linux core, because from the moment we map mckernel text section on linux we will be able to start doing this (and e.g. the hfi driver offloading we have in the ofp branch does this so might call into code that can behave differently if processor_id is -1)
> 
> If this is about do_futex() it might be enough but I still am not sure what you said, please re-explain first

If do_futex () can not be limited, it is impossible to use a scheme to check only when do_futex () is used.

Patch-set: 2
Reviewer: Gerrit User 1015270 <1015270@d5d70762-12d0-45a1-890d-524b12d3f735>