Update patch set 1
Patch Set 1:

Hmm. We can add an ExecStartPost that removes the drop-in as a sort of "self-cleanup" but I'm not comfortable with doing a systemctl daemon-reload there as well to take this into account and not doing it would leave the state dirty which isn't good either; probably not a good idea.

we could just remove it and systemctl daemon-reload within the start script manually after the service has been retarted, but then it might be difficult to tell what's running at a glance, and anyone doing an extra irqbalance restart would screw things up...

Boot time might be a good idea like you said but it's more of a pain to set up, we'd need to install and enable another service with Before=irqbalance.service, which is also kind of meh.
I'll pass for this :p

Patch-set: 1
Reviewer: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>