Update patch set 1
Patch Set 1:

I wonder if this is really what we want, one could simply not be aware of anything else running. How about adding an argument (e.g., -k) and do the kill only when it's specified. Otherwise, it could just report that there are still processes using the device file.

Patch-set: 1
CC: Gerrit User 1015205 <1015205@d5d70762-12d0-45a1-890d-524b12d3f735>