DO NOT MERGE/POC: rus page fault: insert with large page when possible
Due to the way pages are handled on linux, we normally populate an
mckernel large page with multiple small linux pages in the rus page
fault handler. linux transparent huge page do not aggregate these.

Linux huge page handling is a huge pain and there are no helper to
manually "merge" a bunch of pages properly, a real fix would be to
allocate the ihk memory from the start as huge page with mmap() in
ihkconfig and pass that memory to ihk, but as a proof of concept we can
just fill in the page table entry ourselves... As long as we clean it up
properly.

Notes:
 - the cleanup is done when forwarding the exit_group syscall.
There is nowhere we can hook on before the vma cleanup unfortunately;
the vm_operations_struct has a .close() handler but it is called after
the page table is cleaned up for that range so too late for our purpose,
and similarily the /dev/mcos0 fd is closed after the vmas are closed.

That means we will never be able to cleanup properly (e.g. crashing
processes), and in the case of such a failure linux will BUG()

 - the cleanup does not handle the p4d stuff, because lazy won after
realizing it would not be a proper solution anyway

 - there are plenty of locks missing, probably...

 - There must be some other bug, the mellanox driver fails with EFAULT
with this patch at some point. openmpi+bxi gets stuck on 4MB transfers...

Anyway, not useable state ; patch is being uploaded for archiving
purpose only

Change-Id: I4f35552a8cd8479295e0c9619aed10112645be5a