Update patch set 2
Patch Set 2:

So reran on wallaby with the wrapper script, this fails with a segfault an this in kmsg (why the hell don't we have kmsg on failures in logs?)



[  0]: page_fault_process_memory_range: WARNING: cow on non-struct-page-managed page
[  0]: page_fault_process_memory_range: WARNING: cow on non-struct-page-managed page
[  0]: page_fault_process_memory_range: WARNING: cow on non-struct-page-managed page
[  0]: page_fault_process_memory_range: WARNING: cow on non-struct-page-managed page
[  0]: page_fault_process_memory_range: WARNING: cow on non-struct-page-managed page
[  0]: rusage_check_oom: memory used:21461204992 available:21463302144
[  0]: rusage_check_oom: memory used:21461204992 available:21463302144
[  0]: rusage_check_oom: memory used:21461204992 available:21463302144
[  0]: rusage_check_oom: memory used:21461204992 available:21463302144
[  0]: page_fault_process_memory_range(ffff880180b1be20,2aaac0000000-2aaf80000000 723004,2aad3f6f1000,40000006):cannot allocate copy page. -12


Anyway, it's definitely related to this patch, something we didn't treat as cow is now treated as cow for mem_limits... I had wrongly assumed from the name it had something to do with rlimit but it looks like it's about using as much memory as possible?

Looking at the script it runs in mmap mode on 95% of the reserved memory, in one single big anon mmap... then faults all the pages.
I guess this is now handled as cow because range->memobj is set to the zeroobj?... I'd normally think cow should work as well as non-cow for the zero page but... Not too sure what to do, explicitely check for devobjs in the non-page cow case?

Patch-set: 2
Reviewer: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>