munmap: fix deadlock with remote pagefault on vm range lock
While the comment copied from #986 claims this is safe, that is not
correct: if *another thread* gets a page fault at this time, it will
skip grabbing the lock and process the page fault

If the forwarded syscall finishes at this point and release the lock,
the page fault will be unprotected and could corrupt data structures.

Also, now remote page faults can happen out of syscall forwarding as
well, all the other places taking the memory range locks are window for
deadlocks waiting to happen...

On the other hand, the normal fix of using irq-aware spinlocks is not
possible here: you cannot just disable interrupts while sending linux a
syscall. In the first place munmap should likely not take a spinlock for
such a log time...
So a proper fix would be to first remove all places holding the lock for
a long time (over forwarded syscalls at least), then change all lock
sites for memory_range_lock into irq-disabling locks.

Change-Id: I6d9e68e8f8905b20bb2ccfa72848e04fe6404ab6