Update patch set 1
Patch Set 1:

ah, I almost suggested something like that when we did this but let it slide as most machines have sudo installed with a default config that lets root do anything anyway.

If that was required somewhere, wouldn't you want to do that for mcstop+release as well?

Patch-set: 1
CC: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>