Update patch set 1
Patch Set 1:

Probably not far, just wanted to have something that works on ARM64, one caveat with this solution is that if CONFIG_IRQ_WORK is not compiled in this mechanism won't work, although it seems to me it's compiled in basically by default. I will add a check on that somewhere btw.

Patch-set: 1
Reviewer: Gerrit User 1015205 <1015205@d5d70762-12d0-45a1-890d-524b12d3f735>